+++ wallet-api/wallet-api.cabal
@@ -59,6 +59,7 @@ library
         Ledger.Map.TH
         Ledger.These
         Ledger.These.TH
+        Ledger.Crypto
         Ledger.Slot
         Ledger.Scripts
         Ledger.Types
